xml classes moved to firebase functions. Created functions for testing
